FINDING: ACCT-F20260921 LANE: FINANCIAL
SOURCE-OF-TRUTH: factor.faro_daily_imports / factor.faro_invoice_lines (Neon prod, USMCA), proven at apps/backend/src/data-infra/data-infra.service.ts:202-335 (upsertFaroDailyImportOnClient, the table's only writer) and apps/backend/src/accounting/factor-reconciliation/recon.service.ts:72-313 (importStatement)
I QUERIED:       live bypass_rls=lucia reads of factor.faro_daily_imports (1 row), factor.faro_invoice_lines (34 rows), audit.audit_events for loads 13579/13615, accounting.factoring_reserve_movements vs factoring.reserve_movement row counts, information_schema for uq_faro_invoice_lines_per_import
NOT CHECKED:     the full 89-purchase Faro history beyond the single 2026-09-04 statement; invoice minting; reserve-register posting; the 8 cross-entity JE pairs — all named, not attempted, in docs/reconciliation/2026-09-21-round27-28-faro-invoices-handoff.md

ROOT CAUSE: (1) factor.faro_daily_imports' single row (statement 2026-09-04) had header totals ($151,740.00/51 invoices) inconsistent with its own child factor.faro_invoice_lines (34 rows/$104,050.00) because its raw_payload carried a shape {po,inv,date,debtor,escrow_reserve_cents} the app's only writer (upsertFaroDailyImportOnClient) never produces — something wrote directly to the header outside the app. Owner-confirmed (Round 28) the header total was CORRECT (Faro truth); the 34-line ledger is the incomplete side (17 lines never resolved to a USMCA invoice, no error raised). A same-session first attempt shrank the header to match the incomplete ledger — wrong direction, reverted live before this commit, full correction trail left on the row. (2) apps/backend/src/accounting/factor-reconciliation/{recon.service.ts,routes.ts} was fully built and tested but registerFactorReconciliationRoutes was never imported/called in index.ts — every route 404'd, factor.reconciliation_runs had 0 rows. (3) Wiring it live surfaced a real bug: `-$4` on an untyped pg param in the missing_on_statement INSERT throws `operator is not unique: - unknown` (42725) on any advanced-but-unstated invoice. (4) DuplicateVendorsBanner.tsx:135's QBO-id predicate was correct given its target (a QBO-vendor-merge form requiring synced qbo_vendor_id, 0/618 USMCA vendors have one) — the real defect was that the correct, already-built, generic vendor-merge primitive (/api/v1/vendors/:id/flag-duplicate + /merge) had zero frontend callers anywhere.

FIX: Reverted the header to its original, owner-confirmed-correct totals via a live UPDATE computed from the untouched, correct 34-line child table (not hand-typed) plus a full audit note — deliberately did not re-run the app's reimport path because factor.faro_invoice_lines' uq_faro_invoice_lines_per_import unique index is non-partial (does not exclude superseded rows) and throws 23505 on any reimport of a previously-seen invoice_number (live-reproduced, filed, not fixed here — needs a migration). Registered registerFactorReconciliationRoutes in index.ts. Fixed the untyped-operator bug (`-($4::bigint)`). Wired DuplicateVendorsBanner's "Merge these" directly to flag-duplicate+merge with an explicit Keep-A/Keep-B confirm (no automatic survivor pick on a hard-to-reverse merge), added apps/frontend/src/api/factoring.ts client functions. Full extract of the real Faro/invoice source data (9 sheets, ~500 rows) committed to docs/reconciliation/ so the next block doesn't re-derive it.

DOD-A: PASS
DOD-B: N/A
DOD-C: PASS
DOD-D: N/A
DOD-E: PASS
VERIFY-1: PASS
VERIFY-2: N/A
VERIFY-3: PASS
VERIFY-4: PASS
VERIFY-5: N/A
VERIFY-6: N/A
VERIFY-7: PASS
VERIFY-8: UNVERIFIED — no live Chrome click this PR (banner covered by 5 passing RTL tests instead; no user-facing screen changed beyond the banner)
MODULE_PROGRESS: accounting 38 of 40
ITEMS_TOUCHED: ACCT-F20260921
MIGRATE: N/A

GUARD: N/A this commit (the loud-fail import guard from Round 28 step 2 is explicitly NOT built here — see REMAINING; existing test suites apps/backend/src/accounting/factor-reconciliation/__tests__/*.ts and apps/frontend DuplicateVendorsBanner.test.tsx cover the code that did change)
LIVE PROOF: apps/frontend `npm run typecheck` exit 0; apps/backend `npx tsc -b` exit 0; `npx vitest run src/accounting/factor-reconciliation/__tests__/ src/factoring/faro-csv-import.test.ts` 26/26 passing; apps/frontend `npx vitest run src/components/factoring/DuplicateVendorsBanner.test.tsx` 5/5 passing; live Neon (br-fancy-credit-akjnd07a, bypass_rls=lucia) before/after: factor.faro_daily_imports.gross_total_cents restored to 15174000 (was corrupted to 10405000 by this session's own reverted first attempt, was 15174000 originally) with advance/reserve/fee totals all matching the original row exactly; factor.reconciliation_items reproduced the 42725 error pre-fix and produced real missing_in_ledger/matched/missing_on_statement rows post-fix (then deleted, since they were computed against the then-incorrect header — not left as a misleading "first run"); accounting.factoring_reserve_movements confirmed 110 rows/$509,447 cents held for USMCA vs factoring.reserve_movement confirmed 0 rows system-wide.
PROVES-IT-WORKS: docs/reconciliation/2026-09-21-round27-28-faro-invoices-handoff.md carries every query run and its result inline.
KEEPS-IT-TRUE: recompute-on-read (header repair was a one-time correction of historical corruption, not an ongoing derivation)
REMAINING: Live=UNVERIFIED until this deploys and the recon route + banner merge action are clicked live; invoice-per-load minting + QP contra-revenue GL line (51 rows extracted, ready), loud-fail Faro-import guard (Round 28 step 2), full 89-purchase Faro historical load (Round 28 step 3, source extracted), reserve-register posting decision (which table — see finding #4 in the handoff doc), 8 cross-entity JE pairs (source extracted), 13579/13615 status-sync bug (root-caused, not fixed), uq_faro_invoice_lines_per_import non-partial-index bug (root-caused, not fixed), daily-close-to-Faro check, bank-matching suggest-only verification, the two Faro-side-only discrepancies ($4,000.00 cash-reserve gap / $705.02 payments gap — for Faro, not us, to explain). None attempted blind — full detail + real source data in this PR's docs/reconciliation/ files.

+++ apps/backend/src/accounting/factor-reconciliation/recon.service.ts
@@ -276,7 +276,7 @@ export async function importStatement(input: {
             tolerance_cents,
             details
           )
-          VALUES ($1::uuid, $2::uuid, $3::uuid, NULL, 'missing_on_statement', 0, $4, -$4, $5, $6::jsonb)
+          VALUES ($1::uuid, $2::uuid, $3::uuid, NULL, 'missing_on_statement', 0, $4::bigint, -($4::bigint), $5, $6::jsonb)
         `,
         [
           runId,
